Update bmp.js
Here I try a hacky workaround of fixing bmp_save() so it works on
resources with references poining to a canvas.

Because the RGBA mapping differs between the ImageData that canvas
understands as opposed to the BMP format, we can't just call
bmp_create_uri() on a resource created primarily for the Canvas,
this will not work.  Right now, each time I spawn a resource, I
store the |target| inside the resource's property |reference| so
that bmp_save() can have access to it while checking these flags.

Is this awful? Perhaps.  But I'm thinking writing a custom remap
function for rearranging the values back and forth is gonna be more
work than I'm willing to put in.

For now this'll do.