Update bmp.js: This fixes an old bug, again.
In set_pixel and get_pixel functions, we floor() the X and Y coords
so when we do the out of bounds check there's no precision to worry
about.  I don't wanna get into much detail but we most definitely
should not use round() here, not only for performance reasons but
so that we hopefully solve the issue in the process.

I'll need to do font tests also because I've encountered this issue
while working with fonts the other day, where a glyph would start
at the end of an image and its last height column would be drawn at
on the left side of the picture.