TODO(oxou): bmp.js:537
This must be addressed, as we may want to create URLs for canvas-based
resources, but right now we can't because the Web standards suck ass.

Due to these unfortunate artificial "limitations" we have to re-write
`bmp_create_uri` and unfortunately introduce a slight performance hit
on every call to it.

This may also be for the better, as that means that we can get rid of
most of the bloat in the `bmp_save` function so maybe this is a win?

It's also important to note that saving the resources afterwards will
always default to BMP because of this RGBA remapping phase, so this
means anything you create in BMP-JS will be saveable to a BMP file
and also loadable by BMP-JS afterwards.

This gives you "immense" capabilities with the workflow of BMP-JS and
overall functionality provided by it.